WIP: fix noRecmopute flag
Based on `generate_plan.sh`, it seems that it is intended to run Melon twice:
the first time with "noRecompute" set to true to alloc based on seq order in the
default graph; the second time, with the default order generated, run with
"noRecompute" set to false to actually run progressive calibration
recomputation.

But this is still a WIP as SegFaults are triggered when generating the schedule.
Shangchen located the problem to be in the `SegmentTree`.