Revert build_topology()
The up_tensors and down_tensors should be static maps of all spatial adjacencies
that ever occur between tensors.

When I fix the SEGFAULT bug, I added logic to dynamically add and delete up/down
relationships which is not the intended usage of up_tensors and down_tensors.

In the remove_tensor() function, there are extra logic to handle if the tensors
being processed have lifetime overlap. So the up_tensors and down_tensors here
only server as a pre-computation to accelerate the search.